% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_breaks_calc.R
\name{map_breaks_calc}
\alias{map_breaks_calc}
\title{Create a choropleth color scheme for maps based upon jenks breaks}
\usage{
map_breaks_calc(map_obj, field, color_vec = "blues")
}
\arguments{
\item{map_obj}{The dataframe/spatial object with a dataframe that needs color assignment}

\item{field}{The dataframe column/vector that will provide the basis for the choropleth map}

\item{color_vec}{The MEDSL color scheme that will be used for the choropleth map. This argument can be one of five options: "blues",
"purples", "reds", "heat" (green to red), and "heat_rev" (red to green). Note that the blue, red, and purple spectrums each go from 
light to dark. If no option is entered, then it defaults to blues.}
}
\value{

}
\description{
This function takes a dataframe (presumably a spatial object data frame) and assigns medsl themed colors based upon natural 5 jenks breaks. 
These colors can then be carried over into a choropleth map, or a population weighted dot map, which is part of the carto_plot function.
}
\examples{
state_obj <- readOGR(wd_shp, "counties") # reading in the state shpfile 
state_obj <- subset(state_obj, STATEFP=="39") #subset by fip code, which you should have memorized for beer trivia purposes 
##read in the pop data from the acs 
acs_counties <- read.csv("county_acs_demos.csv")
acs_counties$Geo_FIPS <- str_pad(acs_counties$Geo_FIPS, width = 5,pad="0",side="left")
state_obj <- merge(state_obj, acs_counties, by.x="GEOID",by.y="Geo_FIPS")
state_obj <- map_breaks_calc(state_obj, state_obj$white_pct,"purples")

}
