% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameCleanR.R
\name{name_cleaner}
\alias{name_cleaner}
\title{Identify the names that are similar and likely matches}
\usage{
name_cleaner(df, name_field, group_fields, max_distance = 0.1)
}
\arguments{
\item{df}{The long dataframe with the election results of interest}

\item{name_field}{The name of the character/string field of interest that should be matched with similar strings}

\item{group_fields}{The name of the field that the name field should be subsetted by in order to constrain which names are compared against
each other (i.e. county, office, dataverse, etc. )}

\item{max_distance}{The 0 - 1 Levenshtein distance that is used to identify matches, where values closer to 0 are more similar, and 1 
completely different. The default value is 0.1, the recommended}
}
\value{

}
\description{
This function takes the long datasets for other election information and computes the Herfindahl index measure of competition so that 
one can compare it to the raw precinct results
}
\examples{
house_all <- read.csv("2016-precinct-house.csv")
name_match <- name_cleaner(house_all, "candidate", "county", max_distance=0.05) )

}
