% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scytlCleanR.R
\name{scytl_cleaner}
\alias{scytl_cleaner}
\title{Compile and clean the elections data from Scytl websites}
\usage{
scytl_cleaner(file_list1, wd1, wd2, xlsx = TRUE)
}
\arguments{
\item{file_list1}{The vector of file names within the working directory. Can easily be accomplished with the list.files(wd1) cmd}

\item{wd1}{The string of the working directory where the scytl data are saved as excel files by county}

\item{wd2}{The string of the working directory where the long precinct results by county will be saved}

\item{xlsx}{True or false statement as to whether the excel files are in xlsx documents.}
}
\value{

}
\description{
This function takes the wide excel datasets downloaded from Scytl within a working directory and transforms them into long dataframes
by county, with temporary data saved into a secondary working directory.
}
\examples{
ky_wd1 <- "F:/MEDSL/2018-precincts/KY/scytl"
ky_wd2 <- "F:/MEDSL/2018-precincts/KY/cleaned_data"
ky_scytl_list <- list.files(ky_wd1)
scytl_cleaner(ky_scytl_list,ky_wd1,ky_wd2, xlsx = TRUE)
}
