% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scytlCompilR.R
\name{scytl_compiler}
\alias{scytl_compiler}
\title{Compile the scytl results by county into a a single long data frame}
\usage{
scytl_compiler(file_list, wd1)
}
\arguments{
\item{wd1}{The string of the working directory where the scytl_cleaner function saved the cleaned data}

\item{file_list1}{The vector of file names within the working directory. Can easily be accomplished with the list.files(wd) cmd}
}
\value{

}
\description{
This function takes the long by county precinct results from the scytl_cleaner function in a working directory, then binds them into a 
single long dataframe
}
\examples{
ky_wd1 <- "F:/MEDSL/2018-precincts/KY/scytl"
ky_wd2 <- "F:/MEDSL/2018-precincts/KY/cleaned_data"
ky_scytl_list <- list.files(ky_wd1)
scytl_cleaner(ky_scytl_list,ky_wd1,ky_wd2, xlsx = TRUE)
ky_cleaned_list <- list.files(ky_wd2)
ky_precincts_all <- scytl_compiler(ky_cleaned_list,ky_wd2)

}
